Create docs solution and rename `Terraprisma.Ssg` to `Terraprisma.Docs.SSG`.
