+++ docs/generator/Terraprisma.Docs.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Terraprisma.Docs.SSG", "Terraprisma.Docs.SSG\Terraprisma.Docs.SSG.csproj", "{493F036C-63F0-4451-B0E4-D3B85C3F6D92}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{493F036C-63F0-4451-B0E4-D3B85C3F6D92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{493F036C-63F0-4451-B0E4-D3B85C3F6D92}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{493F036C-63F0-4451-B0E4-D3B85C3F6D92}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{493F036C-63F0-4451-B0E4-D3B85C3F6D92}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
